+++ src/apps/chifra/internal/transactions/handle_uniq.go
@@ -31,7 +31,7 @@ func (opts *TransactionsOptions) HandleUniq() (err error, disp bool) {
 					// Address:          mon.Address,
 					BlockNumber:      app.BlockNumber,
 					TransactionIndex: app.TransactionIndex,
-					Timestamp:        ts,
+					Timestamp:        int64(ts),
 					Date:             gostradamus.FromUnixTimestamp(int64(ts)).String(),
 				}
 			}
